+++ core/src/test/java/ai/timefold/solver/core/testdomain/multivar/basic/TestdataMultivarIncrementalScoreCalculator.java
@@ -1,4 +1,4 @@
-package ai.timefold.solver.core.testdomain.multivar;
+package ai.timefold.solver.core.testdomain.multivar.basic;
 
 import java.util.Collection;
 import java.util.Collections;
